added token verifycation for put/delete req

+++ pages/api/products/[id].js
@@ -2,7 +2,8 @@ import dbConnect from "../../../utils/dbConnect";
 import Product from "../../../models/Product"
 
 export default async function handler(req, res) {
-    const { method, query: { id } } = req;
+    const { method, cookies,  query: { id } } = req;
+    const token = cookies.token
 
     await dbConnect()
 
@@ -16,6 +17,9 @@ export default async function handler(req, res) {
           }
           break
         case 'PUT':
+          if(!token || token !== process.env.TOKEN){
+            return res.status(401).json('not authenticated')
+          }
           try {
             const product = await Product.findByIdAndUpdate(id, req.body, {
               new: true,
@@ -26,6 +30,9 @@ export default async function handler(req, res) {
           }
           break
         case 'DELETE':
+          if(!token || token !== process.env.TOKEN){
+            return res.status(401).json('not authenticated')
+          }
           try {
             await Product.findByIdAndDelete(id)
             res.status(200).json("product has been deleted")
